-> ex 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -2,7 +2,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-
+const userExtractor = require('../utils/middleware').userExtractor
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -22,27 +22,22 @@ blogsRouter.get('/:id', async (request, response, next) => {
 })
 
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
 
-blogsRouter.post('/', async (request, response, next) => {
+
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   const body = request.body
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
 
-  if (!user) {
-    return response.status(400).json({ error: 'userId missing or not valid' })
-  }
+  // if (!user) {
+  //   return response.status(400).json({ error: 'userId missing or not valid' })
+  // }
 
-
+  const user = request.user
+  console.log('user', user)
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -60,7 +55,19 @@ blogsRouter.post('/', async (request, response, next) => {
 
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
+  
+  // authorize user
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+  
   await Blog.findByIdAndDelete(request.params.id)
     
       response.status(204).end()
@@ -68,7 +75,7 @@ blogsRouter.delete('/:id', async (request, response, next) => {
   next()
 })
 
-blogsRouter.put('/:id', async (request, response, next) => {
+blogsRouter.put('/:id', userExtractor, async (request, response, next) => {
   const { title, author, url, likes } = request.body
   
   // authorize user
